% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkXy.R
\name{checkXy}
\alias{checkXy}
\title{Check input values.}
\usage{
checkXy(X, y = NULL, map = NULL, covar = NULL)
}
\arguments{
\item{X}{scaled genotype matrix}

\item{y}{phenotype vector}

\item{map}{map of markers}

\item{covar}{covariance matrix}
}
\value{
Nothing if everything is ok. Error or warning otherwise.
}
\description{
Check input values.
}

