% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gensim.R
\name{gensim}
\alias{gensim}
\title{Genetic similarity matrix for scaled genotype X.}
\usage{
gensim(X, map, method = c("LMM", "LOCO", "POMOLOCO"))
}
\arguments{
\item{X}{scaled genotype matrix}

\item{map}{map of markers}

\item{method}{for which method is gensim. matrix calculated}
}
\value{
a square matrix
}
\description{
Genetic similarity matrix for scaled genotype X.
}

