% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heritability.R
\name{heritability}
\alias{heritability}
\title{Estimate heritability.}
\usage{
heritability(y, G, covar = NULL)
}
\arguments{
\item{y}{phenotype vector}

\item{G}{genetic similarity matrix}

\item{covar}{matrix with covariates}
}
\value{
numeric
}
\description{
Estimate heritability.
}

