% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scan.lm.R
\name{scan.lm}
\alias{scan.lm}
\title{One Locus Scan for Linear Model.}
\usage{
scan.lm(y, X, map, covar = NULL)
}
\arguments{
\item{y}{phenotype vector}

\item{X}{scaled genotype matrix}

\item{map}{map of markers}

\item{covar}{matrix with covariates}
}
\value{
R/rql scanone object with LOD scores
}
\description{
One Locus Scan for Linear Model.
}

