% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scan.lmm.R
\name{scan.lmm}
\alias{scan.lmm}
\title{One Locus Scan for Linear Mixed Model.}
\usage{
scan.lmm(y, X, map, covar = NULL, G = NULL)
}
\arguments{
\item{y}{phenotype vector}

\item{X}{scaled genotype matrix}

\item{map}{map of markers}

\item{covar}{matrix with covariates}

\item{G}{genetic similarity matrix}
}
\value{
R/rql scanone object with LOD scores
}
\description{
One Locus Scan for Linear Mixed Model.
}

