% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scan.loco.R
\name{scan.loco}
\alias{scan.loco}
\title{One Locus Scan for for Leave One Chromosome Out Method.}
\usage{
scan.loco(y, X, map, covar = NULL, Glist = NULL)
}
\arguments{
\item{y}{phenotype vector}

\item{X}{scaled genotype matrix}

\item{map}{map of markers}

\item{covar}{matrix with covariates}

\item{Glist}{list of genetic similarity matrices}
}
\value{
R/rql scanone object with LOD scores
}
\description{
One Locus Scan for for Leave One Chromosome Out Method.
}

