% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scan.pomo.R
\name{scan.pomo}
\alias{scan.pomo}
\title{One Locus Scan for Project One Marker Out Method.}
\usage{
scan.pomo(y, X, map, covar = NULL, G = NULL, faster = TRUE)
}
\arguments{
\item{y}{phenotype vector}

\item{X}{scaled genotype matrix}

\item{map}{map of markers}

\item{covar}{matrix with covariates}

\item{G}{genetic similarity matrix}

\item{faster}{overall estimate of heritability?}
}
\value{
R/rql scanone object with LOD scores
}
\description{
One Locus Scan for Project One Marker Out Method.
}

