% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scan.pomoloco.R
\name{scan.pomoloco}
\alias{scan.pomoloco}
\title{One Locus Scan for Project One Marker Out Method.}
\usage{
scan.pomoloco(y, X, map, covar = NULL, Glist = NULL, faster = TRUE)
}
\arguments{
\item{y}{phenotype vector}

\item{X}{scaled genotype matrix}

\item{map}{map of markers}

\item{covar}{matrix with covariates}

\item{Glist}{list of genetic similarity matrices}

\item{faster}{overall estimate of heritability?}
}
\value{
R/rql scanone object with LOD scores
}
\description{
One Locus Scan for Project One Marker Out Method.
}

