% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scanoneTemplate.R
\name{scanoneTemplate}
\alias{scanoneTemplate}
\title{Create R/qtl scanone object}
\usage{
scanoneTemplate(map)
}
\arguments{
\item{map}{map of markers}
}
\description{
Create R/qtl scanone object
}

