% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim.geno.f2.R
\name{sim.geno.f2}
\alias{sim.geno.f2}
\title{Simulate F2 cross genotypes}
\usage{
sim.geno.f2(n.ind, chrlen = 100, nchr = 20, nmar = 30)
}
\arguments{
\item{n.ind}{number of animals}

\item{chrlen}{length of chromosomes}

\item{nchr}{number of chromosomes}

\item{nmar}{number of markers per chromosome}
}
\value{
list with genotype probabilities and map of markers
}
\description{
Simulate F2 cross genotypes
}

