% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim.pheno.R
\name{sim.pheno}
\alias{sim.pheno}
\title{Simulate phenotype with the given background heritability, add QTLs.}
\usage{
sim.pheno(X, G, h2, markers = NULL, effects = NULL)
}
\arguments{
\item{X}{matrix with covariates}

\item{h2}{background heritability}

\item{markers}{which markers}

\item{y}{phenotype vector}
}
\value{
numeric vector
}
\description{
Simulate phenotype with the given background heritability, add QTLs.
}

